Update LoadingBar.vue
Hi there,

I noticed that the .loader-progress-indicator in the LoadingBar. vue component lacks a dynamic width property based on the progress prop.  Adding width: "${progress}%" to the indicator’s style is necessary to ensure it accurately reflects the loading progress.

This addition will improve the user’s perception of the loading process and provide a more engaging visual experience.  Could you please implement this change?
Thanks!